// frontend/app/src/app/model/predmet.model.ts
export class Predmet {
    naziv: string;
    smer: string;
    sifra: string;
    brSmer: number;
    semestar: number;
    student: Array<Object>;
    nastavnik: Array<Object>;
    profesorP: Array<String>;
    profesorV: Array<String>;
    obavestenja: Array<Object>;
    ispit: Array<Object>;
    predavanja: Array<Object>;
    vezbe: Array<Object>;
    projekat: Array<Object>;
    informacije: Array<Object>;
    laboratorija: Array<Object>;
    grupa: Array<Object>;
}